src/backend/app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
aiocoap==0.4.7
paho-mqtt==2.0.0
pydantic==2.6.0
pydantic[email]
python-multipart==0.0.9
